% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{FieldValidation}
\alias{FieldValidation}
\title{FieldValidatio class}
\description{
Provide validator for certain relevant fields of data that needs to be send to the API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FieldValidation-valid_sex}{\code{FieldValidation$valid_sex()}}
\item \href{#method-FieldValidation-valid_bool_vector}{\code{FieldValidation$valid_bool_vector()}}
\item \href{#method-FieldValidation-validate_datetime}{\code{FieldValidation$validate_datetime()}}
\item \href{#method-FieldValidation-check_na_fields}{\code{FieldValidation$check_na_fields()}}
\item \href{#method-FieldValidation-clone}{\code{FieldValidation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldValidation-valid_sex"></a>}}
\if{latex}{\out{\hypertarget{method-FieldValidation-valid_sex}{}}}
\subsection{Method \code{valid_sex()}}{
Validate sex column of a dataframe  as positive integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldValidation$valid_sex(df, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe where the values of the field are}

\item{\code{field}}{field name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldValidation-valid_bool_vector"></a>}}
\if{latex}{\out{\hypertarget{method-FieldValidation-valid_bool_vector}{}}}
\subsection{Method \code{valid_bool_vector()}}{
Validate all values of a column are indeed boolean and not equivalents
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldValidation$valid_bool_vector(df, field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe where the values of the field are}

\item{\code{field}}{field name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldValidation-validate_datetime"></a>}}
\if{latex}{\out{\hypertarget{method-FieldValidation-validate_datetime}{}}}
\subsection{Method \code{validate_datetime()}}{
Validate all values of a column are either datetime in ymd_hms format or a date in ymd format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldValidation$validate_datetime(df, field, is_datetime = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe where the values of the field are}

\item{\code{field}}{field name}

\item{\code{is_datetime}}{boolean flag indicating whether to check datetime or date}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldValidation-check_na_fields"></a>}}
\if{latex}{\out{\hypertarget{method-FieldValidation-check_na_fields}{}}}
\subsection{Method \code{check_na_fields()}}{
Check whether there are NA values in some columns of a given dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldValidation$check_na_fields(df, col_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe}

\item{\code{col_names}}{character vector with the columnnames}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FieldValidation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FieldValidation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FieldValidation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
