% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_endpoint.R
\name{HttpClientJsonBase}
\alias{HttpClientJsonBase}
\title{HttpClientJsonBase class}
\description{
it makes simple atomic CRUD requests to a particular end_point
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HttpClientJsonBase-new}{\code{HttpClientJsonBase$new()}}
\item \href{#method-HttpClientJsonBase-build_request_object}{\code{HttpClientJsonBase$build_request_object()}}
\item \href{#method-HttpClientJsonBase-prepare_df_to_json_conversion}{\code{HttpClientJsonBase$prepare_df_to_json_conversion()}}
\item \href{#method-HttpClientJsonBase-error_handler}{\code{HttpClientJsonBase$error_handler()}}
\item \href{#method-HttpClientJsonBase-get}{\code{HttpClientJsonBase$get()}}
\item \href{#method-HttpClientJsonBase-get_dataframe}{\code{HttpClientJsonBase$get_dataframe()}}
\item \href{#method-HttpClientJsonBase-post_dataframe}{\code{HttpClientJsonBase$post_dataframe()}}
\item \href{#method-HttpClientJsonBase-post_object}{\code{HttpClientJsonBase$post_object()}}
\item \href{#method-HttpClientJsonBase-delete}{\code{HttpClientJsonBase$delete()}}
\item \href{#method-HttpClientJsonBase-clone}{\code{HttpClientJsonBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$new(base_url, end_point)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-build_request_object"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-build_request_object}{}}}
\subsection{Method \code{build_request_object()}}{
Build the http request to a particular endpoint as expected by the HTTP protocol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$build_request_object()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-prepare_df_to_json_conversion"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-prepare_df_to_json_conversion}{}}}
\subsection{Method \code{prepare_df_to_json_conversion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$prepare_df_to_json_conversion(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-error_handler"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-error_handler}{}}}
\subsection{Method \code{error_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$error_handler(e, method = "post")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-get"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-get_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-get_dataframe}{}}}
\subsection{Method \code{get_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$get_dataframe(parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-post_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-post_dataframe}{}}}
\subsection{Method \code{post_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$post_dataframe(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-post_object"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-post_object}{}}}
\subsection{Method \code{post_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$post_object(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-delete"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$delete(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HttpClientJsonBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HttpClientJsonBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HttpClientJsonBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
