% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iim_api_consumer.R
\name{IIMApiClient}
\alias{IIMApiClient}
\title{IIMApiClient class}
\description{
Provides the end-to-end queries required to update all details about the live acoustic telemetry system in place
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IIMApiClient-new}{\code{IIMApiClient$new()}}
\item \href{#method-IIMApiClient-add_sensor_type}{\code{IIMApiClient$add_sensor_type()}}
\item \href{#method-IIMApiClient-post_data_to_api}{\code{IIMApiClient$post_data_to_api()}}
\item \href{#method-IIMApiClient-get_from_api}{\code{IIMApiClient$get_from_api()}}
\item \href{#method-IIMApiClient-get_from_api_as_dataframe}{\code{IIMApiClient$get_from_api_as_dataframe()}}
\item \href{#method-IIMApiClient-clone}{\code{IIMApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$new(api_url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-add_sensor_type"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-add_sensor_type}{}}}
\subsection{Method \code{add_sensor_type()}}{
Add a sensor type to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$add_sensor_type(df, end_point = "sensortypes/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{single-column dataframe with as many variables as required, i.e., DEPTH, TEMPERATURE.}

\item{\code{end_point}}{string with the name of the end point being queried. Default "sensortypes/"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-post_data_to_api"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-post_data_to_api}{}}}
\subsection{Method \code{post_data_to_api()}}{
Post a nested list to the database through an API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$post_data_to_api(obj, end_point)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{a nested list resulting from applying \link{NestedListBuilder}}

\item{\code{end_point}}{string with the name of the end point being queried. Default "tags_with_sensors/"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-get_from_api"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-get_from_api}{}}}
\subsection{Method \code{get_from_api()}}{
Make a request to the API and return the data in a json object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$get_from_api(end_point)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end_point}}{string with the name of the end point being queried. Default "tags_with_sensors/"}

\item{\code{obj}}{a nested list resulting from applying \link{NestedListBuilder}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-get_from_api_as_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-get_from_api_as_dataframe}{}}}
\subsection{Method \code{get_from_api_as_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$get_from_api_as_dataframe(end_point, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IIMApiClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IIMApiClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IIMApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
