% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_list_builder.R
\name{NestedListBuilder}
\alias{NestedListBuilder}
\title{NestedListBuilder class}
\value{
a nested list where each element at the root level contains all details for each tag and its sensor
details.

a nested list where each element at the root level contains all details for each tagged individual.
}
\description{
Convert dataframes into nested lists according to each required case
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NestedListBuilder-build_tag_with_sensor_list}{\code{NestedListBuilder$build_tag_with_sensor_list()}}
\item \href{#method-NestedListBuilder-build_tagged_individual_list}{\code{NestedListBuilder$build_tagged_individual_list()}}
\item \href{#method-NestedListBuilder-clone}{\code{NestedListBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedListBuilder-build_tag_with_sensor_list"></a>}}
\if{latex}{\out{\hypertarget{method-NestedListBuilder-build_tag_with_sensor_list}{}}}
\subsection{Method \code{build_tag_with_sensor_list()}}{
Given a dataframe with the tag and sensor details, it returns 4-level nested list containing all details
about each tag along with their sensor details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedListBuilder$build_tag_with_sensor_list(
  df,
  tag_prefix,
  sensor_prefix,
  root_field_names
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe with all tag and sensor details}

\item{\code{tag_prefix}}{string with the expected tag field prefix}

\item{\code{sensor_prefix}}{string with the expected sensor field prefix}

\item{\code{root_field_names}}{string vector with the names of the outer element of the built list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedListBuilder-build_tagged_individual_list"></a>}}
\if{latex}{\out{\hypertarget{method-NestedListBuilder-build_tagged_individual_list}{}}}
\subsection{Method \code{build_tagged_individual_list()}}{
Given a dataframe with the individual, tag and sensor and tagged_individual details it returns 3-level nested list
containing all such details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedListBuilder$build_tagged_individual_list(
  df,
  ind_prefix,
  tag_prefix,
  sensor_prefix,
  tagged_ind_prefix,
  root_field_names
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe with all tag and sensor details}

\item{\code{ind_prefix}}{string with the expected individual field prefix}

\item{\code{tag_prefix}}{string with the expected tag field prefix}

\item{\code{sensor_prefix}}{string with the expected sensor field prefix}

\item{\code{tagged_ind_prefix}}{string with the expected tagged_individual field prefix}

\item{\code{root_field_names}}{string vector with the names of the outer element of the built list in the strict order
of (individual, tag, sensor)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NestedListBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NestedListBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NestedListBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
