% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templo_api_client.R
\name{TemploApiClient}
\alias{TemploApiClient}
\title{TemploApi class}
\value{
TRUE or FALSE
}
\description{
API client to consume TEMPLO's API
Add a tag with a sensor details to the database.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TemploApiClient-new}{\code{TemploApiClient$new()}}
\item \href{#method-TemploApiClient-add_tag_with_sensor}{\code{TemploApiClient$add_tag_with_sensor()}}
\item \href{#method-TemploApiClient-add_species}{\code{TemploApiClient$add_species()}}
\item \href{#method-TemploApiClient-add_tagged_individuals}{\code{TemploApiClient$add_tagged_individuals()}}
\item \href{#method-TemploApiClient-add_single_record_and_fetch_id}{\code{TemploApiClient$add_single_record_and_fetch_id()}}
\item \href{#method-TemploApiClient-get_sensor_types}{\code{TemploApiClient$get_sensor_types()}}
\item \href{#method-TemploApiClient-get_tagged_individuals}{\code{TemploApiClient$get_tagged_individuals()}}
\item \href{#method-TemploApiClient-get_environmental_readings}{\code{TemploApiClient$get_environmental_readings()}}
\item \href{#method-TemploApiClient-get_dataframe_end_point}{\code{TemploApiClient$get_dataframe_end_point()}}
\item \href{#method-TemploApiClient-clone}{\code{TemploApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$new(api_url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-add_tag_with_sensor"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-add_tag_with_sensor}{}}}
\subsection{Method \code{add_tag_with_sensor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$add_tag_with_sensor(
  df,
  tag_prefix,
  sensor_prefix,
  field_names,
  end_point = "tags_with_sensors/"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{end_point}}{string with the name of the end point being queried. Default "tags_with_sensors/"}

\item{\code{obj}}{a nested list resulting from applying \link{NestedListBuilder}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-add_species"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-add_species}{}}}
\subsection{Method \code{add_species()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$add_species(df, end_point = "species/")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-add_tagged_individuals"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-add_tagged_individuals}{}}}
\subsection{Method \code{add_tagged_individuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$add_tagged_individuals(
  df,
  individual_prefix,
  tag_prefix,
  sensor_prefix,
  tagged_individual_prefix,
  field_names,
  end_point = "tagged_individuals/"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-add_single_record_and_fetch_id"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-add_single_record_and_fetch_id}{}}}
\subsection{Method \code{add_single_record_and_fetch_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$add_single_record_and_fetch_id(end_point, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-get_sensor_types"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-get_sensor_types}{}}}
\subsection{Method \code{get_sensor_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$get_sensor_types(end_point = "sensortypes/")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-get_tagged_individuals"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-get_tagged_individuals}{}}}
\subsection{Method \code{get_tagged_individuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$get_tagged_individuals(end_point = "tagged_individuals/")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-get_environmental_readings"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-get_environmental_readings}{}}}
\subsection{Method \code{get_environmental_readings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$get_environmental_readings(
  api_url,
  end_point = "environmental_readings/data_exports/",
  from_date,
  to_date
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-get_dataframe_end_point"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-get_dataframe_end_point}{}}}
\subsection{Method \code{get_dataframe_end_point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$get_dataframe_end_point(api_url, end_point, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TemploApiClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TemploApiClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TemploApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
